requirements.txt
torch~=2.4.1
ultralytics~=8.2.92
transformers~=4.44.2
cv2
requests~=2.32.2
ibm-cos-sdk
flask~=3.0.3
numpy~=1.24.4
Flask-SocketIO
Flask-Caching
firebase-admin
flask-cors
google-cloud-firestore
opencv-python~=4.10.0.84
torchvision~=0.19.1
pillow~=10.4.0


mediapipe~=0.10.11
tensorflow~=2.13.0
keras~=2.13.1
